requirements.txt
albumentations==1.1.0
h5py==2.10.0
imantics==0.1.10
numpy==1.18.2
Keras-Preprocessing==1.1.2
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.32
pandas==1.1.5
patchify
seaborn==0.11.1
segmentation-models
tensorflow-gpu==2.4.0